Refactor warnings
