+++ domenetjenester/iay/src/main/java/no/nav/foreldrepenger/abakus/domene/iay/AktørArbeidEntitet.java
@@ -41,7 +41,7 @@ public class AktørArbeidEntitet extends BaseEntitet implements AktørArbeid, In
 
     @ManyToOne(optional = false)
     @JoinColumn(name = "inntekt_arbeid_ytelser_id", nullable = false, updatable = false)
-    private InntektArbeidYtelseAggregatEntitet inntektArbeidYtelser;
+    private InntektArbeidYtelseAggregat inntektArbeidYtelser;
 
     @OneToMany(mappedBy = "aktørArbeid")
     @ChangeTracked
@@ -92,7 +92,7 @@ public class AktørArbeidEntitet extends BaseEntitet implements AktørArbeid, In
         this.yrkesaktiviter = new LinkedHashSet<>();
     }
 
-    void setInntektArbeidYtelser(InntektArbeidYtelseAggregatEntitet inntektArbeidYtelser) {
+    void setInntektArbeidYtelser(InntektArbeidYtelseAggregat inntektArbeidYtelser) {
         this.inntektArbeidYtelser = inntektArbeidYtelser;
     }
